+++ src/app/modules/search-parameter/autocomplete-parameter-value.component.ts
@@ -11,7 +11,7 @@ import {
   Self,
   ViewChild
 } from '@angular/core';
-import { FormControl, NgControl } from '@angular/forms';
+import { AbstractControl, FormControl, NgControl } from '@angular/forms';
 import { BaseControlValueAccessor } from '../base-control-value-accessor';
 import Def from 'autocomplete-lhc';
 import { MatFormFieldControl } from '@angular/material/form-field';
@@ -56,7 +56,6 @@ export interface Lookup {
 export class AutocompleteParameterValueComponent
   extends BaseControlValueAccessor<AutocompleteParameterValue>
   implements
-    OnChanges,
     OnDestroy,
     AfterViewInit,
     MatFormFieldControl<AutocompleteParameterValue> {
@@ -87,6 +86,20 @@ export class AutocompleteParameterValueComponent
       this.input?.nativeElement.className.indexOf('invalid') >= 0 ||
       (formControl && this.errorStateMatcher.isErrorState(formControl, null))
     );
+  }
+
+  /**
+   * Whether the control has required validator (Implemented as part of MatFormFieldControl)
+   */
+  get required(): boolean {
+    const validator = this.ngControl?.control.validator;
+    if (validator) {
+      const exampleResult = validator({} as AbstractControl);
+      if (exampleResult && exampleResult.required) {
+        return true;
+      }
+    }
+    return false;
   }
 
   /**
@@ -175,7 +188,6 @@ export class AutocompleteParameterValueComponent
    */
   readonly disabled: boolean = false;
   readonly id: string;
-  readonly required = false;
 
   /**
    * Returns EV id from a DbGap variable API response
@@ -220,12 +232,6 @@ export class AutocompleteParameterValueComponent
   onContainerClick(event: MouseEvent): void {
     if (!this.focused) {
       document.getElementById(this.inputId).focus();
-    }
-  }
-
-  ngOnChanges(): void {
-    if (this.acInstance) {
-      this.setupAutocomplete();
     }
   }
 
@@ -318,7 +324,11 @@ export class AutocompleteParameterValueComponent
     return new Def.Autocompleter.Prefetch(
       this.inputId,
       this.options.map((o) => o.display),
-      { maxSelect: '*', codes: this.options.map((o) => o.code) }
+      {
+        maxSelect: '*',
+        codes: this.options.map((o) => o.code),
+        matchListValue: true
+      }
     );
   }
 
@@ -760,6 +770,12 @@ export class AutocompleteParameterValueComponent
       codes: [],
       items: []
     };
+    if (this.acInstance) {
+      this.currentData.items.forEach((item, index) => {
+        this.acInstance.storeSelectedItem(item, this.currentData.codes[index]);
+        this.acInstance.addToSelectedArea(item);
+      });
+    }
   }
 
   /**