+++ src/app/modules/search-parameter/search-parameter.component.spec.ts
@@ -13,7 +13,7 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { MatInputModule } from '@angular/material/input';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { AutocompleteComponent } from '../autocomplete/autocomplete.component';
+import { AutocompleteModule } from '../autocomplete/autocomplete.module';
 
 class Page {
   private fixture: ComponentFixture<SearchParameterComponent>;
@@ -37,7 +37,6 @@ describe('SearchParameterComponent', () => {
       {
         declarations: [
           SearchParameterComponent,
-          MockComponent(AutocompleteComponent),
           MockComponent(ObservationTestValueComponent)
         ],
         imports: [
@@ -48,7 +47,8 @@ describe('SearchParameterComponent', () => {
           MatFormFieldModule,
           MatAutocompleteModule,
           MatInputModule,
-          NoopAnimationsModule
+          NoopAnimationsModule,
+          AutocompleteModule
         ]
       },
       {